Add JPetTaskIOLoopPerSubTask class
It is added to contain the run method from the current JPetTaskIO.
In addition JPetTaskIO becomes an abstract class.
