+++ include/Core/JPetTaskIO/JPetTaskIO.h
@@ -16,14 +16,14 @@
 #ifndef JPETTASKIO_H
 #define JPETTASKIO_H
 
-#include "./JPetProgressBarManager/JPetProgressBarManager.h"
-#include "./JPetTaskInterface/JPetTaskInterface.h"
 #include "./JPetParamManager/JPetParamManager.h"
-#include "./JPetStatistics/JPetStatistics.h"
-#include "./JPetTaskIO/JPetOutputHandler.h"
-#include "./JPetTaskIO/JPetInputHandler.h"
 #include "./JPetParams/JPetParams.h"
+#include "./JPetProgressBarManager/JPetProgressBarManager.h"
+#include "./JPetStatistics/JPetStatistics.h"
 #include "./JPetTask/JPetTask.h"
+#include "./JPetTaskIO/JPetInputHandler.h"
+#include "./JPetTaskIO/JPetOutputHandler.h"
+#include "./JPetTaskInterface/JPetTaskInterface.h"
 #include <memory>
 #include <string>
 
@@ -34,12 +34,10 @@ class JPetStatistics;
 /**
  * @brief Helper structure to encapsulate some fields
  */
-struct TaskIOFileInfo {
-  TaskIOFileInfo(
-    const std::string& inType, const std::string& outType,
-    const std::string& outFullPath, bool resetOutPath):
-    fInFileType(inType), fOutFileType(outType),
-    fOutFileFullPath(outFullPath), fResetOutputPath(resetOutPath) {};
+struct TaskIOFileInfo
+{
+  TaskIOFileInfo(const std::string& inType, const std::string& outType, const std::string& outFullPath, bool resetOutPath)
+      : fInFileType(inType), fOutFileType(outType), fOutFileFullPath(outFullPath), fResetOutputPath(resetOutPath){};
   std::string fInFileType;
   std::string fOutFileType;
   std::string fOutFileFullPath;
@@ -48,31 +46,27 @@ struct TaskIOFileInfo {
 
 /**
  * @brief Class representing computing task with input/output operations.
- * In the current implementation the single entry that is read by the reader
- * corresponds to a JPetTimeWindow object.
+ * It is not meant to be used directly, rather used as specialized classes
+ * that must provide an implementation of the method run().
  */
-class JPetTaskIO: public JPetTask
+class JPetTaskIO : public JPetTask
 {
 public:
   JPetTaskIO(const char* name = "", const char* in_file_type = "", const char* out_file_type = "");
   virtual ~JPetTaskIO();
   virtual bool init(const JPetParams& inOptions) override;
-  virtual bool run(const JPetDataInterface& inData) override;
+  virtual bool run(const JPetDataInterface& inData) = 0;
   virtual bool terminate(JPetParams& outOptions) override;
   virtual void addSubTask(std::unique_ptr<JPetTaskInterface> subTask) override;
   void displayProgressBar(std::string taskName, int currentEventNumber, int numberOfEvents) const;
-  virtual JPetParams mergeWithExtraParams(
-    const JPetParams& originalParams, const JPetParams& extraParams
-  ) const ;
+  virtual JPetParams mergeWithExtraParams(const JPetParams& originalParams, const JPetParams& extraParams) const;
   void setParams(const JPetParams& opts);
   JPetParams getParams() const;
   bool isOutput() const;
   bool isInput() const;
 
 protected:
-  virtual std::tuple<bool, std::string, std::string, bool> setInputAndOutputFile(
-    const jpet_options_tools::OptsStrAny options
-  ) const;
+  virtual std::tuple<bool, std::string, std::string, bool> setInputAndOutputFile(const jpet_options_tools::OptsStrAny options) const;
   virtual bool createInputObjects(const char* inputFilename);
   virtual bool createOutputObjects(const char* outputFilename);
   const JPetParamBank& getParamBank();
