+++ src/Core/JPetTaskIO/JPetTaskIO.cpp
@@ -84,85 +84,85 @@ bool JPetTaskIO::init(const JPetParams& params)
   return true;
 }
 
-bool JPetTaskIO::run(const JPetDataInterface&)
-{
-  using namespace jpet_options_tools;
-  if (fSubTasks.empty())
-  {
-    ERROR("No subTask set");
-    return false;
-  }
-  if (isInput())
-  {
-    if (!fInputHandler)
-    {
-      ERROR("No inputHandler set");
-      return false;
-    }
-  }
-  for (const auto& pTask : fSubTasks)
-  {
-    auto subTaskName = pTask->getName();
-    bool isOK = pTask->init(fParams);
+// bool JPetTaskIO::run(const JPetDataInterface&)
+//{
+// using namespace jpet_options_tools;
+// if (fSubTasks.empty())
+//{
+// ERROR("No subTask set");
+// return false;
+//}
+// if (isInput())
+//{
+// if (!fInputHandler)
+//{
+// ERROR("No inputHandler set");
+// return false;
+//}
+//}
+// for (const auto& pTask : fSubTasks)
+//{
+// auto subTaskName = pTask->getName();
+// bool isOK = pTask->init(fParams);
 
-    if (!isOK)
-    {
-      WARNING("In init() of:" + subTaskName + ". run()  and terminate() of this task will be skipped.");
-      continue;
-    }
+// if (!isOK)
+//{
+// WARNING("In init() of:" + subTaskName + ". run()  and terminate() of this task will be skipped.");
+// continue;
+//}
 
-    if (isInput())
-    {
-      assert(fInputHandler);
-      bool isProgressBarOn = isProgressBar(fParams.getOptions());
-      isOK = fInputHandler->setEntryRange(fParams.getOptions());
-      if (!isOK)
-      {
-        ERROR("Some error occured in setEntryRange");
-        return false;
-      }
-      auto lastEvent = fInputHandler->getLastEntryNumber();
-      assert(lastEvent >= 0);
-      do
-      {
-        if (isProgressBarOn)
-        {
-          displayProgressBar(subTaskName, fInputHandler->getCurrentEntryNumber(), lastEvent);
-        }
-        JPetData event(fInputHandler->getEntry());
-        isOK = pTask->run(event);
-        if (!isOK)
-        {
-          ERROR("In run() of:" + subTaskName + ". ");
-          return false;
-        }
-        if (isOutput())
-        {
-          if (!fOutputHandler->writeEventToFile(pTask.get()))
-          {
-            ERROR("Some problems occured, while writing the event to file.");
-            return false;
-          }
-        }
-      } while (fInputHandler->nextEntry());
-    }
-    else
-    {
-      JPetDataInterface dummyEvent;
-      pTask->run(dummyEvent);
-    }
+// if (isInput())
+//{
+// assert(fInputHandler);
+// bool isProgressBarOn = isProgressBar(fParams.getOptions());
+// isOK = fInputHandler->setEntryRange(fParams.getOptions());
+// if (!isOK)
+//{
+// ERROR("Some error occured in setEntryRange");
+// return false;
+//}
+// auto lastEvent = fInputHandler->getLastEntryNumber();
+// assert(lastEvent >= 0);
+// do
+//{
+// if (isProgressBarOn)
+//{
+// displayProgressBar(subTaskName, fInputHandler->getCurrentEntryNumber(), lastEvent);
+//}
+// JPetData event(fInputHandler->getEntry());
+// isOK = pTask->run(event);
+// if (!isOK)
+//{
+// ERROR("In run() of:" + subTaskName + ". ");
+// return false;
+//}
+// if (isOutput())
+//{
+// if (!fOutputHandler->writeEventToFile(pTask.get()))
+//{
+// ERROR("Some problems occured, while writing the event to file.");
+// return false;
+//}
+//}
+//} while (fInputHandler->nextEntry());
+//}
+// else
+//{
+// JPetDataInterface dummyEvent;
+// pTask->run(dummyEvent);
+//}
 
-    JPetParams subTaskParams;
-    isOK = pTask->terminate(subTaskParams);
-    if (!isOK)
-    {
-      ERROR("In terminate() of:" + subTaskName + ". ");
-      return false;
-    }
-    fParams = mergeWithExtraParams(fParams, subTaskParams);
-  }
-  return true;
-}
+// JPetParams subTaskParams;
+// isOK = pTask->terminate(subTaskParams);
+// if (!isOK)
+//{
+// ERROR("In terminate() of:" + subTaskName + ". ");
+// return false;
+//}
+// fParams = mergeWithExtraParams(fParams, subTaskParams);
+//}
+// return true;
+//}
 
 bool JPetTaskIO::terminate(JPetParams& output_params)
 {
